Refactor NeutronDiffraction into diffraction.py
